requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
plotly>=5.17.0
numpy>=1.24.0